requirements.txt
pandas
requests
beautifulsoup4
psycopg2-binary
openpyxl
lxml